requirements.txt
matplotlib==3.1.2
numpy==1.18.1
pandas==1.0.3
scipy==1.4.1
folium==0.10.1 
gpxpy==1.4.0
progressbar2==3.47.0
garminconnect==0.1.15